apply_newbook_handler / book_handler / book_mybook_handler

+++ Server/routes/book_mybook_handler.js
@@ -9,7 +9,11 @@ function loadBorrowedBook(req, res){
     var con = DB_handler.connectDB();
     var query = 'SELECT *, DATEDIFF(CURDATE(), a.b_due_date) diff FROM t_book a INNER JOIN t_book_rental b ON a.b_id=b.br_book_id where b.br_user="' + req.session.passport.user.id + '"';
     con.query(query, function(err, response){
-        res.send(response);
+        if(err){
+            console.log('DB select ERROR "book_mybook_handler.js -> loadBorrowedBook');
+        }else{
+            res.send(response);
+        }
         DB_handler.disconnectDB(con);
     });
 }
@@ -18,7 +22,12 @@ function loadReservedBook(req, res){
     var con = DB_handler.connectDB();
     var query = 'SELECT * FROM t_book a INNER JOIN t_book_reserve b ON a.b_id=b.bre_book_id where b.bre_user="' + req.session.passport.user.id + '"';
     con.query(query, function(err, response){
-        res.send(response);
+        if(err){
+            console.log('DB select ERROR "book_mybook_handler.js -> loadReservedBook');
+        }else{
+            res.send(response);
+
+        }
         DB_handler.disconnectDB(con);
     });
 }
@@ -27,7 +36,12 @@ function loadAppliedBook(req, res){
     var con = DB_handler.connectDB();
     var query = 'SELECT * FROM t_book_apply where ba_user="' + req.session.passport.user.id + '"';
     con.query(query, function(err, response){
-        res.send(response);
+        if(err){
+            console.log('DB select ERROR "book_mybook_handler.js -> loadAppliedBook');
+        }else{
+            res.send(response);
+
+        }
         DB_handler.disconnectDB(con);
     });
 }
@@ -42,31 +56,33 @@ function turninBook(req, res){
     con.query(query, function(err, response){
         if(err){
             res.send('failed');
-            throw err
-        }
-        var due_date = new Date(req.body.due_date);
-        var t = new Date(today);
-        var diff = (t.getTime() - due_date.getTime()) / (1000 * 60 * 60 * 24);
-
-        if(diff > 0){
-            util.send(req.session.passport.user.id, '책 반납일 미준수', util.pushContents.b_turnin, function(err, data){
-                if(err){
-                    console.log(err);
-                }else{
-                    console.log(data);
-                }
-            });
-            var query_imposeDuty = 'update t_user set u_bad_duty_point=u_bad_duty_point+' + diff + ' where u_id="' + req.session.passport.user.id + '";';
-            query_imposeDuty += 'insert into t_duty_point_history SET date="' + today + '", receive_user="' + req.session.passport.user.id + '", send_user="' + response[3][0].u_id + '", mode=1, point=' + diff + ', reason="책 반납일 미준수"';
-            con.query(query_imposeDuty, function(err2, response2){
-                if(err2){
-                    res.send('failed');
-                    throw err2
-                }
-                res.send(diff);
-            })
+            console.log('DB query ERROR "book_mybook_handler.js -> turninBook');
         }else{
-            res.send('success');
+            var due_date = new Date(req.body.due_date);
+            var t = new Date(today);
+            var diff = (t.getTime() - due_date.getTime()) / (1000 * 60 * 60 * 24);
+
+            if(diff > 0){
+                util.send(req.session.passport.user.id, '책 반납일 미준수', util.pushContents.b_turnin, function(err2, data){
+                    if(err2){
+                        console.log(err2);
+                    }else{
+                        console.log(data);
+                    }
+                });
+                var query_imposeDuty = 'update t_user set u_bad_duty_point=u_bad_duty_point+' + diff + ' where u_id="' + req.session.passport.user.id + '";';
+                query_imposeDuty += 'insert into t_duty_point_history SET date="' + today + '", receive_user="' + req.session.passport.user.id + '", send_user="' + response[3][0].u_id + '", mode=1, point=' + diff + ', reason="책 반납일 미준수"';
+                con.query(query_imposeDuty, function(err3, response3){
+                    if(err3){
+                        res.send('failed');
+                        console.log('DB query2 ERROR "book_mybook_handler.js -> turninBook');
+                    }else{
+                        res.send(diff);
+                    }
+                })
+            }else{
+                res.send('success');
+            }
         }
     });
     if(req.body.reserved_cnt > 0)   push2Subscriber(con, req.session.passport.user.id, req.body.book_id);
@@ -79,16 +95,21 @@ function postponeBook(req, res){
     var query2 = 'update t_book_rental set br_extension_cnt=br_extension_cnt+1 where br_id=' + req.body.rental_id + ';';
     query2 += 'update t_book set b_due_date=ADDDATE(b_due_date, 14) where b_id="' + req.body.book_id + '"';
     con.query(query, function(err, response){
-        if(response[0].br_extenstion_cnt === 1 || response[0].b_reserved_cnt != 0){
-             res.send('연장할 수 없습니다');
+        if(err){
+            console.log('DB select ERROR "book_mybook_handler.js -> postponeBook');
         }else{
-            con.query(query2, function(err2, response2){
-                if(err2){
-                    res.send('failed');
-                    throw err2
-                }
-                res.send('success');
-            });
+            if(response[0].br_extenstion_cnt === 1 || response[0].b_reserved_cnt != 0){
+                res.send('연장할 수 없습니다');
+            }else{
+                con.query(query2, function(err2, response2){
+                    if(err2){
+                        res.send('failed');
+                        console.log('DB update ERROR "book_mybook_handler.js -> postponeBook');
+                    }else{
+                        res.send('success');
+                    }
+                });
+            }
         }
         DB_handler.disconnectDB(con);
     });
@@ -104,9 +125,11 @@ function missingBook(req, res){
     con.query(query, function(err, response){
         if(err){
             res.send('failed');
-            throw err
+            console.log('DB query ERROR "book_mybook_handler.js -> missingBook');
+        }else{
+            res.send('success');
+
         }
-        res.send('success');
         DB_handler.disconnectDB(con);
     });
 }
@@ -118,18 +141,21 @@ function cancelReservation(req, res){
         con.query(query, function(err, response){
             if(err){
                 res.send('failed');
-                throw err
+                console.log('DB select ERROR "book_mybook_handler.js -> cancelReservation');
+            }else{
+                var query1 = 'update t_book_reserve set bre_myturn=bre_myturn-1 where bre_myturn>'+ response[0].bre_myturn + ';';
+                query1 += 'update t_book set b_reserved_cnt=b_reserved_cnt-1 where b_id="' + req.body.book_id + '";';
+                query1 += 'delete from t_book_reserve where bre_id="' + req.body.reserve_id + '"';
+                con.query(query1, function(err2, response2){
+                    if(err2){
+                        res.send('failed');
+                        console.log('DB query ERROR "book_mybook_handler.js -> cancelReservation');
+                    }else{
+                        res.send('success');
+
+                    }
+                });
             }
-            var query1 = 'update t_book_reserve set bre_myturn=bre_myturn-1 where bre_myturn>'+ response[0].bre_myturn + ';';
-            query1 += 'update t_book set b_reserved_cnt=b_reserved_cnt-1 where b_id="' + req.body.book_id + '";';
-            query1 += 'delete from t_book_reserve where bre_id="' + req.body.reserve_id + '"';
-            con.query(query1, function(err2, response2){
-                if(err2){
-                    res.send('failed');
-                    throw err2
-                }
-                res.send('success');
-            });
         });
         DB_handler.disconnectDB(con);
     }
@@ -139,11 +165,14 @@ function cancelAppliedbook(req, res){
     var con = DB_handler.connectDB();
     var query = 'delete from t_book_apply where ba_id="' + req.body.apply_id + '"';
     con.query(query, function(err, response){
-        if(err){
+        if(err) {
             res.send('failed');
-            throw err
+            console.log('DB query ERROR "book_mybook_handler.js -> cancelAppliedbook');
+
+        }else{
+
+            res.send('success');
         }
-        res.send('success');
         DB_handler.disconnectDB(con);
     });
 }
@@ -153,27 +182,50 @@ function push2Subscriber(userId, book_id){
     var due_date = getDate(new Date(), 14);
     var query1 = 'select * FROM t_user a INNER JOIN t_book_reserve b ON a.u_id=b.bre_user where b.bre_book_id="' + book_id + '" and b.bre_myturn=1';
     con.query(query1, function(err, response){
-        util.send(response[0].u_id, '대여 알림', util.pushContents.b_borrow, function(err_push, data){
-            if(err_push){
-                console.log(err_push);
-            }else{
-                console.log(data);
-            }
-        });
-        var query2 = 'update t_book set b_state=1, b_due_date="' + due_date + '", b_rental_username="' + response[0].u_name + '", b_reserved_cnt=b_reserved_cnt-1 where b_id="' + book_id + '"';
-        con.query(query2);
-        var query3 = 'insert into t_book_rental SET ?';
-        var queryData = {
-            br_user: response[0].u_id,
-            br_book_id: book_id,
-            br_rental_date: getDate(new Date(), 0)
-        };
-        con.query(query3, queryData);
+        if(err){
+            console.log('DB select ERROR "book_mybook_handler.js -> push2Subscriber');
+        }else{
+            util.send(response[0].u_id, '대여 알림', util.pushContents.b_borrow, function(err_push, data){
+                if(err_push){
+                    console.log(err_push);
+                }else{
+                    console.log(data);
+                }
+            });
+            var query2 = 'update t_book set b_state=1, b_due_date="' + due_date + '", b_rental_username="' + response[0].u_name + '", b_reserved_cnt=b_reserved_cnt-1 where b_id="' + book_id + '"';
+            con.query(query2, function(err2, response2){
+                if(err2){
+                    console.log('DB update ERROR "book_mybook_handler.js -> push2Subscriber');
+                }else{
+                    var query3 = 'insert into t_book_rental SET ?';
+                    var queryData = {
+                        br_user: response[0].u_id,
+                        br_book_id: book_id,
+                        br_rental_date: getDate(new Date(), 0)
+                    };
+                    con.query(query3, queryData, function(err3, response3){
+                        if(err3){
+                            console.log('DB insert ERROR "book_mybook_handler.js -> push2Subscriber');
+                        }
+                    });
+                }
+            });
+        }
     });
     var query4 = 'update t_book_reserve set bre_myturn=bre_myturn-1 where bre_book_id="' + book_id + '"';
-    con.query(query4);
-    var query5 = 'delete from t_book_reserve where bre_myturn=0 and bre_book_id="' + book_id + '"';
-    con.query(query5);
+    con.query(query4, function(err4, response4){
+        if(err4){
+            console.log('DB update ERROR "book_mybook_handler.js -> push2Subscriber');
+        }else{
+            var query5 = 'delete from t_book_reserve where bre_myturn=0 and bre_book_id="' + book_id + '"';
+            con.query(query5, function(err5, response5){
+                if(err5){
+                    console.log('DB delete ERROR "book_mybook_handler.js -> push2Subscriber');
+                }
+            });
+        }
+    });
+
     DB_handler.disconnectDB(con);
 }
